api_python/requirements.txt
fastapi
uvicorn
pydantic
requests
boto3
matplotlib
sgkit
duckdb
pandas
psycopg2-binary
tqdm